SOC-7942: misc: bcm-vk: Add 32bit seq num and fd when dumping stuck msgs
In order to provide more info when messages are found stuck inside
the driver, a new 32-bit seq_num is added, which fixes the limitation
of msg_id having only 4k max.  The later gets wrapped around easily
when multiple sessions (eg 8 streams) are launched simulataneously.
The idx of the ctx is also printed to distinguish one "fopen"
of the device node from another.

Change-Id: I52447fb33db7ddf4bba341f32b072e8ceda7c379
Reviewed-on: http://gerrit-ccxsw.broadcom.net/102232
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>