SOC-10532: misc:bcm-vk: Modify boot2 download to reduce transfer time
Modify boot2 download to reduce transfer time.
The bcm-vk driver is modified to use the pull method of data transfer
from host to card. The pull method requires the card to initiate the
dma transfer from the buffer shared by the bcm-vk to the card.

The bcm-vk driver updates the dma buffer with the boot2 file contents
and sends notification to the card to indicate availability of data.

The card initiates the transfer using its internal DMA engine that
fetches the data from dma buffer and transfers to card's internal
memory. The card sends ACK status to indicate the completion of the
data transfer from host dma buffer to card's internal memory.

After receiving the ACK status from card, the driver updates the
next block from the file. The driver continues to update the dma
buffer, until the firmware update complete status is set by the card.

Change-Id: I29e054c6e3b41290f31520bec885a7bea6e928fd
Reviewed-on: http://gerrit-ccxsw.broadcom.net/104538
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>