SOC-10585: soc: bcm: iproc: Allow panic to be disabled from IDM
An IDM timeout is usually fatal, non-recoverable error on the bus. Upon
detection of the IDM timeout, kernel panic should be the default
behavior of the IDM driver. However, there are cases that user would
want to disable the kernel panic after the detection of IDM timeout,
for debugging and experiment purposes.

This change allows user to disable the default kernel panic behavior
upon detection of IDM timeout via a sysfs entry.

Change-Id: Iec31248d98538fd719d9a9506f39948256d90da9
Signed-off-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/104528
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Rayagonda Kokatanur <rayagonda.kokatanur@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>