SOC-10585: soc: iproc: do not clear memory if signature is detected
During IDM ELOG driver probe, the memory to store the IDM events is
always cleared. In some cases for example when crash kernel is enabled,
it causes the logged events to be cleared when the driver is probed
again in the crash kernel. Change the logic to only clear the memory
when there's no ELOG signature detected.

Change-Id: I54f9e2b891cf2211fb91a37502697437e1691852
Signed-off-by: Ray Jui <ray.jui@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/105189
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>