SOC-10686: misc: bcm-vk: Add spinlock protection for msg_id bitmap clear
When a msg_id is allocated, the corresponding bitmap is set and will
be unset when the msg_id response is returned.  However, in the clearing
case, the call to bitmap_clear is not protected with msg_id_lock.
Since the bitmap operations are read/modify/write, if somehow a set and
clear happen on the same word (where the bits are allocated) but with
different threads, it will result in corruption, causing losing of msg_ids.
Add protection also in the bitmap clearing call.

Change-Id: I5a80db418358088b36b74ab21234fc4981caf096
Reviewed-on: http://gerrit-ccxsw.broadcom.net/105480
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>