SOC-10461: misc: bcm-vk: Add check for limit in ring buffer comm
Separate the static ring buffer info from the dynanmic rd/wr idx.
This allows runtime checking against the idx range in case the PCIe
interface goes down.  In addition, this would be optimized a bit as
the size/offset do not need to be fetched each time from the PCIe
interface.

Change-Id: Icf176195aedde2fd8c9eba3870b9157ca8e9f3da
Reviewed-on: http://gerrit-ccxsw.broadcom.net/105614
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>