SOC-10532: misc:bcm-vk:Abstract read firmware into buf API
Abstract read firmwware into buf API.
The request_firmware_into_buf API needs to support the offset and
pread_flags to handle partial download of large files.
The current kernel version does not support the partial file read,
and so the additional args offset and pread_flags are not
available.
When the bcm-vk driver is built locally in ubuntu machine, the
bcm-vk driver needs to be modified to compile without the offset
and pread_flags. In order to minimize the changes required in driver
for different kernel versions, a definition is added to abstract the
API "request_firmware_into_buf" definition. All kernel builds that do
not support the use of offset and pread_flags has to define
the flag "CONFIG_REQ_FW_INTO_BUF_PRIV' in the build.

The abstracted version of the firmware read API
"request_firmware_into_buf_priv" does not handle partial file
transfer.

Change-Id: I96e939b465d2515491684751eb2c2bd397bbff61
Reviewed-on: http://gerrit-ccxsw.broadcom.net/105755
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Mohamed Ismail Abdul Packir Mohamed <mohamed-ismail.abdul@broadcom.com>
Tested-by: Mohamed Ismail Abdul Packir Mohamed <mohamed-ismail.abdul@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>