SOC-10532 misc: bcm-vk: Modify boot2 download to use pull method
Modify boot2 download to use pull method.
The bcm-vk driver is modified to use the pull method of data transfer
for downloading boot2 image from host to card. The pull method of
data transfer is relatively faster than the push method and it also
make the transfer time independent of the host environment.

The pull method requires the card to initiate the dma transfer from
the buffer shared by the bcm-vk driver to the card. The boot2 image is
updated into the dma buffer in chunks and card is notified to pull the
image from the dma buffer.
The card initiates the transfer using its internal DMA engine that
fetches the data from dma buffer and transfers to card's internal memory.
Card sends acknowledgment through the BAR mapped memory, to indicate the
completion of the data transfer.
Host continues to fill the dma buffer with the next chunk(s) from the
boot2 file, until all the image has been received by the card.

Change-Id: I2bb4996d18a3797f728b74e4b4b653093726cf7d
Reviewed-on: http://gerrit-ccxsw.broadcom.net/105937
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>