SOC-10449: devlink: add support for reporter recovery completion
It is possible that a reporter recovery completion do not finish
successfully when recovery is triggered via
devlink_health_reporter_recover as recovery could be processed in
different context. In such scenario an error is returned by driver when
recover hook is invoked and successful recovery completion is
intimated later.
Expose devlink recover done API to update recovery stats.

Change-Id: I3c6f8053d56e9b413becf709efcd8049c3d29568
Signed-off-by: Vikas Gupta <vikas.gupta@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/106796
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Vasundhara Volam <vasundhara-v.volam@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>