SOC-10461: misc: bcm-vk: Add thresholds for ECC/UECC alerts
When ECC/UECC happen, keep track of the occurrence and only if the
total occurrence crosses the specific threshold that a dmesg will
be output.  This is a bit different from other alerts where it is
only an ON/OFF operation.
Also change and simplify the naming of the alert defines.

Change-Id: Id5fe5bf4a0368d713aea4d8e42766c6e6726b6ce
Reviewed-on: http://gerrit-ccxsw.broadcom.net/107521
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>