SOC-10951: misc: bcm-vk: Delay workqueue destroy on module removal
For some kernel, when the PCIe interrupts are freed at module removal,
there seems to be bogus interrrupts being triggered.  Since the
workqueue is removed earlier, this will the system to reboot. This
happens only with some older Centos.  In addition, also add checking
of drv access and if off, no need to schedule job.

Change-Id: Ib7106731980cdeb4148c7eea7b4a7ba6bdde73e3
Reviewed-on: http://gerrit-ccxsw.broadcom.net/108854
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>