SOC-11166 bnxt_en: Add missing devlink health reporters for VFs.
The VF driver also needs to create the health reporters since
VFs are also involved in firmware reset and recovery.  Modify
bnxt_dl_register() and bnxt_dl_unregister() so that they can
be called by the VFs to register/unregister devlink.  Only the PF
will register the devlink parameters.  With devlink registered,
we can now create the health reporters on the VFs.

Change-Id: I576926319825eee01d794f32b6c9695658acdeca
Fixes: 6763c779c2d8 ("bnxt_en: Add new FW devlink_health_reporter")
Signed-off-by: Vasundhara Volam <vasundhara-v.volam@broadcom.com>
Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/111870
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>