SOC-10966 iproc_defconfig: add CONFIG_QED to select CONFIG_NET_DEVLINK
add CONFIG_QED to iproc_defconfig to have a driver to select
CONFIG_NET_DEVLINK in the case the BNXT is build out of tree.
Unfortunately kernel has changed the Kconfig rules for CONFIG_NET_DEVLINK
and it can't be chosen in the defconfig file anymore so we need this
workaround.

Change-Id: I4da87fdc9e55458e49cf369afb68175f548e7296
Reviewed-on: http://gerrit-ccxsw.broadcom.net/109075
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Tested-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/109284
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>