SOC-11002: bcm_vk_dev.c: fix warning message due to an incompatible type
When compiling the bcm_vk driver, it reports a warning message due to
trying to pass in a size_t variable to a variable of type unsigned int.
However, it does not stop the compilation. The warning message contains
reference to a line in the code that has the word 'Error'. That causes
one of the QA tests to stop running because they stop when they detect
the word error in the output. Fix so QA tests detects real errors only.

Change-Id: Ibcd090a7777c0e56a14a4f7c1567f6f7db5250db
Reviewed-on: http://gerrit-ccxsw.broadcom.net/109487
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>