SOC-11082: misc: bcm-vk: Skip autoload on card with corrupted QSPI
After power-reset, a VK card that has corrupted QSPI will have SRAM
open.  If host driver launches the auto-load to download boot1, it
will fail and retrigger the BROM to run resulting FB_reg being 0x64
(0x64 - SBI not working).
Since BROM/chip has the undesirable "feature" that it will auto-clear
SRAM, all the code will be whipped, probably including the jump vector.
The card now enters a dead state and will not respond to any soft-reset,
which makes it not soft-recoverable.
Therefore, we skip downloading boot1 if the FB_reg is not indicating
that BROM has run to completion.  This would leave the card in the
original SRAM-open, powered up state.

Change-Id: I813202b059eb7f59af2effbd406a1eac98d71a8d
Reviewed-on: http://gerrit-ccxsw.broadcom.net/110713
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>