CTRL-46114: defconfig: Add defconfig for sr2 emulation
Add sr2 emulation defconfig which is reduced version of iproc defconfig.
Many configurations are disable and make it minimal to use for bringup
Linux in emulators. After validate any module or IP, its  configuration
will be added to this file.

This file will be removed once all modules and IPs are validated
in emulation.

Change-Id: I0edcc43275a60af631d1b5f675022e9852dd9a9e
Signed-off-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-on: http://gerrit-ccxsw.broadcom.net/112626
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>