SOC-11379: misc: bcm-vk: reset heartbeat counter everytime on alive
Reset the heartbeat loss count back to 0 everytime a difference
is detected.  Recent usage of the card seems to introduce occasional
heartbeat miss due to statistics being collected for which the logger
on the card is running at same priority as the watchdog kicking thread.
This makes occasional heartbeat count stall possible, and so we have
to reset the counter to 0 whenever a heartbeat diff is detected.

Change-Id: Ib544bbd2b7acf10dceef32ec9ecee9b4f348ac67
Reviewed-on: http://gerrit-ccxsw.broadcom.net/115825
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Jeffery Li <jeffery.li@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>