SOC-11243: misc: bcm-vk: allow time for dma operation during draining
ffmpeg will run asynchronously from the VK card, and it could hit a
situation that a DMA transfer request is shipped but the ffmpeg get
killed.  In this case, the VK card may still hold on to a SGL and may
also step on memory indicated by it.  Need to allow time for the
operation to happen before freeing or closing the fd.

Change-Id: I068d20985a2f9425f0adb384d2dff93933fa2a7c
Reviewed-on: http://gerrit-ccxsw.broadcom.net/124694
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Tested-by: vpx_autobuild status <vpx_autobuild@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Chintan Patel <chintan.patel@broadcom.com>
Tested-by: Chintan Patel <chintan.patel@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>