SOC-11243: misc: brcm-vk: add parmeter to limit burst logging
In a densely populated server with high number of VK cards,
if an ffmpeg session is killed with signal, there will be a
burst of dev level debug logs due to stuck messages or messages
with unknown msg_ids.  This burstiness is worst if
a script is killed where the script has launched n number of
sessions.  The burstiness of logs may overflow the server's
console capacity, resulting in rcu stall detected.
Therefore, a parameter to control such burst logging is
introduced.  The default is to print the first line per-burst.

Change-Id: I6d8e64b62f81f8b10c277449ac8e3dc9e0d3e11c
Reviewed-on: http://gerrit-ccxsw.broadcom.net/125191
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>