SOC-12034: misc: bcm-vk: skip backtrace collection on corrupted ptrs
Do a check of the peerlog pointers against the record collected at
boot2 startup, and if pointers are found corrupted, log a message
and skip backtrace collection.  It is always possible that a sys
fault on a card is due to memory being corrupted and we should do a
sanity check before dumping.

Change-Id: Ibf9c8b4156fa0a72c72ac8561c9f6b5bcbc574ef
Reviewed-on: http://gerrit-ccxsw.broadcom.net/129186
Reviewed-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>