SOC-12291 bcm-vk: misc: Use unique host pid for individual vk streams
In the kernel the pid is analogous to a user-space thread id while the
thread-group ID is analogous to the user-space process id. If user uses
containers or VM, then the pid (get_pid) passed down may not be unique,
as different namespace may overlap.

Convert the internal house-keeping using the unique pid (host), and when
init is passed down from vkil, replaces the value with the unique id
instead of using the user-space one.  This is similar mechanisim as the
unique msg_id host driver creates.

Change-Id: I02ae63818e7d713f3f79873a466921abd620d587
Reviewed-on: https://gerrit-ccxsw.broadcom.net/133819
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: CCXSW Build <ccxsw.build@broadcom.com>
Tested-by: CCXSW Build <ccxsw.build@broadcom.com>
Reviewed-by: Graham Rushton <graham.rushton@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>
Reviewed-on: https://gerrit-ccxsw.broadcom.net/134365
Reviewed-by: CCXSW Coverity
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>