SOC-12006 i2c: add i2c recovery for bcm-iproc based devices
It is possible for the i2c bus to become locked up preventing
communication with devices on the bus. This can occur when
another i2c device fails to be reset correctly. In this case
the SDA line will be held low preventing further communication
on the bus.

This situation can be corrected by sending a series of clock
pulses to allow the blocking device to complete the transaction
and release the bus.

Add the hooks required to allow the existing i2c recovery code
to be used to clear the lock up.

Before the recovery can be performed the device needs to be
configured in the bit-bang mode to allow lines to be
controlled directly by generic i2c recovery mechanism.
Access routines are provided to get and set the lines
state. On completion, the device is returned to
normal operations with initialization.

Based on the patch from Richard Laing.
https://lkml.org/lkml/2019/3/25/1090

Signed-off-by: Richard Laing <richard.laing@alliedtelesis.co.nz>
Signed-off-by: Vladimmir Olovyannikov <vladimir.olovyannikov@broadcom.com>

Change-Id: I274b9604224a7e36d9d7d39e7126ad682050e0ca
Reviewed-on: https://gerrit-ccxsw.broadcom.net/134622
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Vladimir Olovyannikov <vladimir.olovyannikov@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: CCXSW Coverity <ccxswcoverity@broadcom.com>
Tested-by: Manasa Mudireddy <manasa.mudireddy@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>