SOC-12266: misc: bcm-vk: add support for thermal trap alert
Add support of new alert bit to indicate that card has entered
thermal trap state.  This state is entered if the card detects
temperature exceeding its set trap threshold, which is also
signified by a special reason.  In this state, no boot2 is allowed
as DDR is not initialized, and an additional manual reset is
required.

Change-Id: Ib0e7f4dd20eba3d14cadd287a84b3797e3af157f
Reviewed-on: https://gerrit-ccxsw.broadcom.net/135607
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: CCXSW Coverity <ccxswcoverity@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>