SOC-12264: misc: bcm-vk: prevent double killing a job on blk access
If the kernel fails to close a fd which makes the entry still in the
hashed table, the item may be killed again when wrap around happens
on the 24bit namespace pid + during hardware access turning off.
This is only a theoretical corner case, but have got request to
handle that. Therefore, add logic to prevent this.
If an item has been killed for a long time but still persists in
hashed table, move it to an isolated list, for which would be freed
at driver unprobe time.  Previous way of pre-allocating contexts is
replaced by dynamic allocation instead, as collision may occur if
the ctx has been recycled which meets the one that has been delayed
long enough by kernel before closing it.

Change-Id: Iabef4ecc4b66d22b5fd885fda8d12d050e4d2adf
Reviewed-on: https://gerrit-ccxsw.broadcom.net/136265
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Shamez Kurji <shamez.kurji@broadcom.com>
Reviewed-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Desmond Yan <desmond.yan@broadcom.com>
Tested-by: Ccxsw Build <ccxsw.build@broadcom.com>
Reviewed-by: CCXSW Coverity <ccxswcoverity@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>