CTRL-52186: watchdog: sp805_wdt: add support of single stage and two stage modes
Add a new module parameter in watchdog driver to allow user to
select either single stage mode or two stage mode. In two stage mode,
the first interrupt(bark) is handled by a force panic and timeout
is adjusted accordingly. By doing so, the kdump kernel is invoked
and we can collect with core kdump kernel for analysis. In single stage
mode, the first interrupt(bark) is ignored.

Change-Id: I9d6a1af16d43132444fc7fd1ca3a422bb95ec2be
Reviewed-on: https://gerrit-ccxsw.broadcom.net/139012
Reviewed-by: Yuqing Shen <yuqing.shen@broadcom.com>
Reviewed-by: vpx_checkpatch status <vpx_checkpatch@broadcom.com>
Reviewed-by: Scott Branden <scott.branden@broadcom.com>
Reviewed-by: Ray Jui <ray.jui@broadcom.com>
Tested-by: Ray Jui <ray.jui@broadcom.com>